UserService/requirements.txt
django
djangorestframework
django-extensions
python-dotenv
requests
cryptography